+++ src/main/java/it/ccse/uscite/application/facade/dto/input/gestionale/LavorazioneContabileNota_InDTO.java
@@ -3,6 +3,8 @@
  */
 package it.ccse.uscite.application.facade.dto.input.gestionale;
 
+import java.math.BigInteger;
+
 import it.ccse.uscite.application.facade.dto.input.Base_InDTO;
 
 
@@ -17,19 +19,19 @@ public class LavorazioneContabileNota_InDTO extends Base_InDTO  {
 	 */
 	private static final long serialVersionUID = 1L;
 	
-	private Integer idNota;
+	private BigInteger idNota;
 
 	/**
 	 * @return the idNota
 	 */
-	public Integer getIdNota() {
+	public BigInteger getIdNota() {
 		return idNota;
 	}
 
 	/**
 	 * @param idNota the idNota to set
 	 */
-	public void setIdNota(Integer idNota) {
+	public void setIdNota(BigInteger idNota) {
 		this.idNota = idNota;
 	}
