+++ src/main/java/it/ccse/uscite/application/facade/dto/input/gestionale/SearchNote_InDTO.java
@@ -3,6 +3,7 @@
  */
 package it.ccse.uscite.application.facade.dto.input.gestionale;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import it.ccse.uscite.application.facade.dto.input.PageableDTO;
@@ -21,11 +22,11 @@ public class SearchNote_InDTO extends PageableDTO{
 	 */
 	private static final long serialVersionUID = 5375396477621450197L;
 
-	private Integer idNota;
+	private BigInteger idNota;
 	
 	private String owner;
 	
-	private Integer idComitato;
+	private BigInteger idComitato;
 	
 	private Integer numeroNota;
 	
@@ -81,28 +82,28 @@ public class SearchNote_InDTO extends PageableDTO{
 	/**
 	 * @return the idComitato
 	 */
-	public Integer getIdComitato() {
+	public BigInteger getIdComitato() {
 		return idComitato;
 	}
 
 	/**
 	 * @param idComitato the idComitato to set
 	 */
-	public void setIdComitato(Integer idComitato) {
+	public void setIdComitato(BigInteger idComitato) {
 		this.idComitato = idComitato;
 	}
 
 	/**
 	 * @return the idNota
 	 */
-	public Integer getIdNota() {
+	public BigInteger getIdNota() {
 		return idNota;
 	}
 
 	/**
 	 * @param idNota the idNota to set
 	 */
-	public void setIdNota(Integer idNota) {
+	public void setIdNota(BigInteger idNota) {
 		this.idNota = idNota;
 	}
